Shunt server state into a different module
